04_Professionnal__(Days82_to_Day100)/Day92__ColorScan/requirements.txt
Pillow==11.2.1
numpy==2.2.4
scikit-learn==1.6.1
Flask==3.1.0
gunicorn==23.0.0